pyproject.toml
[project]
name = "resume-buliding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "chromadb>=1.5.0",
    "diffusers>=0.36.0",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.129.0",
    "ffmpeg-python>=0.2.0",
    "google-generativeai>=0.8.6",
    "google-search-results>=2.4.2",
    "hf-xet>=1.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-mistralai>=1.1.1",
    "langchain-nomic>=1.0.1",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.8",
    "mysqlclient>=2.2.8",
    "openai-whisper>=20250625",
    "pillow>=11.3.0",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
    "safetensors>=0.7.0",
    "sentence-transformers>=5.2.2",
    "sentencepiece>=0.2.1",
    "tavily-python>=0.7.21",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "wikipedia>=1.4.0",
]
